Made PoseDetectorTasksVision work with two poses
This allows us to do multiplayer games. From two player it will always make sure that the person on the left is blue and red and the person on the right is orange and purple. Also allows code to toggle this so that it doesn't affect other portions of the code